+++ spec/rspec/matchers/built_in/raise_error_spec.rb
@@ -22,14 +22,16 @@ RSpec.describe "expect { ... }.to raise_error" do
     expect { raise StandardError.new, 'boom' }.to raise_error
   end
 
-  it "issues a warning when `nil` is passed for an error class" do
-    expect_warning_with_call_site __FILE__, __LINE__+1, /with a `nil`/
-    expect { raise }.to raise_error(nil)
+  it "raises an exception when `nil` is passed for an error class" do
+    expect {
+      expect { raise }.to raise_error(nil)
+    }.to raise_error(/with a `nil`/)
   end
 
-  it "issues a warning when `nil` is passed for an error class when negated" do
-    expect_warning_with_call_site __FILE__, __LINE__+1, /raise_error\(nil\)/
-    expect { '' }.not_to raise_error(nil)
+  it "raises an exception when `nil` is passed for an error class when negated" do
+    expect {
+      expect { '' }.not_to raise_error(nil)
+    }.to raise_error(/with a `nil`/)
   end
 
   it "issues a warning that does not include current error when it's not present" do